+++ src/AngleSharp/Html/Dom/Internal/HtmlNoScriptElement.cs
@@ -1,4 +1,4 @@
-﻿namespace AngleSharp.Html.Dom
+namespace AngleSharp.Html.Dom
 {
     using AngleSharp.Dom;
     using System;
@@ -9,8 +9,18 @@
     sealed class HtmlNoScriptElement : HtmlElement
     {
         public HtmlNoScriptElement(Document owner, String? prefix = null)
-            : base(owner, TagNames.NoScript, prefix, NodeFlags.Special | NodeFlags.LiteralText)
+            : base(owner, TagNames.NoScript, prefix, GetFlags(owner.Context))
         {
+        }
+
+        private static NodeFlags GetFlags(IBrowsingContext context)
+        {
+            if (context.IsScripting())
+            {
+                return NodeFlags.Special | NodeFlags.LiteralText;
+            }
+
+            return NodeFlags.Special;
         }
     }
 }