+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Released on Sunday, January 8 2023
 - Updated build system to use NUKE instead of CAKE (#1075) @driekus77
 - Fixed `TagClosedWrong` error with nested formatting elements (#1052)
 - Fixed performance issue within deeply nested structures (#1066) @heinrich-ulbricht
+- Fixed decoding of entities in `noscript` elements without scripting (#1070)
 - Fixed setting templte content using `InnerHtml` (#1072)
 - Fixed `TreeWalker` throwing `NullReferenceException` (#1073)
 - Improved `GetItemByIndex` performance (#1050) @egil